+++ SGI/SGI/Models/VentaItem.cs
@@ -10,7 +10,7 @@ namespace SGI.Models
         public int VentaId { get; set; }
         [Required]
         public int ProductoId { get; set; }
-        public decimal? Precio { get; set; } = 0;
+        public Double? Precio { get; set; } = 0;
         [Required]
         public int Cantidad { get; set; } = 1;
         [ForeignKey("VentaId")]
@@ -19,5 +19,5 @@ namespace SGI.Models
         [ForeignKey("ProductId")]
         public Producto? Producto { get; set; }
 
-        public decimal? Subtotal { get; set; } = 0;
+        public Double? Subtotal { get; set; } = 0;
     } }
